JS Insulin Delivery System
A simple system to calculate how much insulin you are required to take based on your current sugar reading, and the amount of food you are eating in carb value. DISCLAIMER This was a project I created while learning JavaScript and all numbers are a default value. It's important that your get your values from your Diabetes Educator if you intend to use this program.